Order OCO

+++ Model/WorkCryptoBirge/Models/AssetBalance.cs
@@ -3,7 +3,7 @@
     public class AssetBalance
     {
         public string Asset { get; set; }
-        public double Free { get; set; }      
-        public double Locked { get; set; }
+        public decimal Free { get; set; }      
+        public decimal Locked { get; set; }
     }
 }
